requirements.txt
# requirements.txt

# Web Server
fastapi
uvicorn
gunicorn
pydantic

# RAG Pipeline
httpx
python-dotenv
numpy
PyMuPDF
python-docx
sentence-transformers
faiss-cpu
openai